Move Reorderable into ui/
The dnd-kit-backed Reorderable primitive is generic UI, not a board-builder
concept. Promoted from features/board-builder/useReorderable.tsx →
ui/Reorderable/Reorderable.tsx. KidSequence and BoardBuilder both consume
it from ui/, removing the last cross-feature import (kid-sequence →
board-builder) along with the queries-vs-mutations split fixed in the
previous commit.

Filename also corrected: it was named after a hook (useReorderable) but
exports a component (Reorderable).
