+++ src/features/kid-sequence/KidSequence.tsx
@@ -1,9 +1,5 @@
 import { type JSX, useState } from 'react';
 
-import {
-  type DragBindings,
-  Reorderable,
-} from '@/features/board-builder/useReorderable';
 import { KidModeLayout } from '@/features/kid-mode/KidModeLayout';
 import { useSetStepIds } from '@/lib/queries/boards';
 import { usePictogramsById } from '@/lib/queries/pictograms';
@@ -11,6 +7,7 @@ import { speakPictogram } from '@/lib/voiceOut';
 import type { Board, Pictogram } from '@/types/domain';
 import { SpeakerIcon } from '@/ui/icons';
 import { PictogramMedia } from '@/ui/PictoTile/PictogramMedia';
+import { type DragBindings, Reorderable } from '@/ui/Reorderable/Reorderable';
 
 import styles from './KidSequence.module.css';
